+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/engine/mynlp/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * MYNLP 中文NLP工具包分词实现<br>
+ * 项目地址：https://github.com/mayabot/mynlp/
+ * 
+ * @author Looly
+ * @since 4.6.5
+ */
+package cn.hutool.extra.tokenizer.engine.mynlp;